backend/requirements.txt
pyyaml==6.0.2
joblib==1.5.1
pandas==2.3.1
fastapi==0.116.1
lightgbm==4.6.0
uvicorn==0.35.0
python-dotenv==1.1.1
imbalanced-learn==0.12.4
scikit-learn==1.4.2
xgboost==3.0.2
catboost==1.2.8
matplotlib==3.10.3
seaborn==0.13.2
pydantic==2.11.7